updated somehtings
