+++ ollama.log
@@ -173,3 +173,4 @@ time=2025-07-24T18:43:35.610-04:00 level=INFO source=server.go:637 msg="llama ru
 [GIN] 2025/07/24 - 18:44:11 | 500 | 14.909932957s |       127.0.0.1 | POST     "/api/generate"
 [GIN] 2025/07/24 - 18:55:51 | 200 |    1.993928ms |       127.0.0.1 | GET      "/api/tags"
 [GIN] 2025/07/24 - 19:01:49 | 200 |     462.151µs |       127.0.0.1 | GET      "/api/tags"
+[GIN] 2025/07/28 - 11:16:04 | 200 |   27.174064ms |       127.0.0.1 | GET      "/api/tags"
